# backend/接口文档.openapi (2).json
{
  "openapi": "3.1.0",
  "info": {
    "title": "默认模块",
    "description": "校园设备管理系统",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/repair/orders/page": {
      "get": {
        "summary": "分页查询本人维修工单",
        "deprecated": false,
        "description": "",
        "operationId": "分页查询本人维修工单_repair_orders_page_get",
        "tags": [
          "维修端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页条数",
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "维修工单状态编码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "维修工单状态编码",
              "title": "Status"
            }
          },
          {
            "name": "equipmentName",
            "in": "query",
            "description": "设备名称",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备名称",
              "title": "Equipmentname"
            }
          },
          {
            "name": "repairUserId",
            "in": "query",
            "description": "维修人员 ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "维修人员 ID",
              "title": "Repairuserid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Page_RepairOrderPageOut__"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/repair/orders/{repairOrderId}": {
      "get": {
        "summary": "查看本人维修工单详情",
        "deprecated": false,
        "description": "",
        "operationId": "查看本人维修工单详情_repair_orders__repairOrderId__get",
        "tags": [
          "维修端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "repairOrderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Repairorderid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_RepairOrderOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/repair/orders/{repairOrderId}/accept": {
      "post": {
        "summary": "接收维修工单",
        "deprecated": false,
        "description": "",
        "operationId": "接收维修工单_repair_orders__repairOrderId__accept_post",
        "tags": [
          "维修端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "repairOrderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Repairorderid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_RepairOrderActionOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/repair/orders/{repairOrderId}/start": {
      "post": {
        "summary": "开始维修工单",
        "deprecated": false,
        "description": "",
        "operationId": "开始维修工单_repair_orders__repairOrderId__start_post",
        "tags": [
          "维修端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "repairOrderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Repairorderid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_RepairOrderActionOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/repair/orders/{repairOrderId}/completion": {
      "post": {
        "summary": "提交维修结果",
        "deprecated": false,
        "description": "",
        "operationId": "提交维修结果_repair_orders__repairOrderId__completion_post",
        "tags": [
          "维修端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "repairOrderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Repairorderid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairOrderCompletionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_RepairOrderCompletionOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/repair-orders/page": {
      "get": {
        "summary": "分页查询全部维修工单",
        "deprecated": false,
        "description": "",
        "operationId": "分页查询全部维修工单_admin_repair_orders_page_get",
        "tags": [
          "管理端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页条数",
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "维修工单状态编码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "维修工单状态编码",
              "title": "Status"
            }
          },
          {
            "name": "equipmentName",
            "in": "query",
            "description": "设备名称",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备名称",
              "title": "Equipmentname"
            }
          },
          {
            "name": "repairUserId",
            "in": "query",
            "description": "维修人员 ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "维修人员 ID",
              "title": "Repairuserid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Page_RepairOrderPageOut__"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/repair-orders/{repairOrderId}": {
      "get": {
        "summary": "查看维修工单详情",
        "deprecated": false,
        "description": "",
        "operationId": "查看维修工单详情_admin_repair_orders__repairOrderId__get",
        "tags": [
          "管理端/维修工单相关"
        ],
        "parameters": [
          {
            "name": "repairOrderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Repairorderid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_RepairOrderOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context",
            "properties": {}
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Page_RepairOrderPageOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RepairOrderPageOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[RepairOrderPageOut]"
      },
...
      "RepairOrderActionOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "维修工单状态"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "维修工单 ID"
          },
          "equipmentId": {
            "type": "integer",
            "title": "Equipmentid",
            "description": "设备 ID"
          },
          "equipmentStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipmentstatus",
            "description": "设备状态"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "Updatetime",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "status",
          "id",
          "equipmentId",
          "updateTime"
        ],
        "title": "RepairOrderActionOut"
      },
...
      "RepairOrderCompletionIn": {
        "properties": {
          "resultStatus": {
            "type": "string",
            "pattern": "^(repaired|unrepairable)$",
            "title": "Resultstatus",
            "description": "维修结果状态"
          },
          "faultCause": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Faultcause",
            "description": "故障原因"
          },
          "repairProcess": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Repairprocess",
            "description": "维修过程"
          },
          "repairResult": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Repairresult",
            "description": "维修结果"
          },
          "beforeImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 9,
            "minItems": 1,
            "title": "Beforeimages",
            "description": "维修前图片地址"
          },
          "afterImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 9,
            "minItems": 1,
            "title": "Afterimages",
            "description": "维修后图片地址"
          }
        },
        "type": "object",
        "required": [
          "resultStatus",
          "faultCause",
          "repairProcess",
          "repairResult",
          "beforeImages",
          "afterImages"
        ],
        "title": "RepairOrderCompletionIn"
      },
      "RepairOrderCompletionOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "维修工单状态"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "维修工单 ID"
          },
          "equipmentId": {
            "type": "integer",
            "title": "Equipmentid",
            "description": "设备 ID"
          },
          "equipmentStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipmentstatus",
            "description": "设备状态"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "Updatetime",
            "description": "更新时间"
          },
          "faultCause": {
            "type": "string",
            "title": "Faultcause",
            "description": "故障原因"
          },
          "repairProcess": {
            "type": "string",
            "title": "Repairprocess",
            "description": "维修过程"
          },
          "repairResult": {
            "type": "string",
            "title": "Repairresult",
            "description": "维修结果"
          },
          "beforeImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Beforeimages",
            "description": "维修前图片地址"
          },
          "afterImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Afterimages",
            "description": "维修后图片地址"
          },
          "completionTime": {
            "type": "string",
            "format": "date-time",
            "title": "Completiontime",
            "description": "提交维修完成时间"
          }
        },
        "type": "object",
        "required": [
          "status",
          "id",
          "equipmentId",
          "updateTime",
          "faultCause",
          "repairProcess",
          "repairResult",
          "completionTime"
        ],
        "title": "RepairOrderCompletionOut"
      },
      "RepairOrderOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "维修工单状态"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "维修工单 ID"
          },
          "repairReportId": {
            "type": "integer",
            "title": "Repairreportid",
            "description": "报修记录 ID"
          },
          "equipmentId": {
            "type": "integer",
            "title": "Equipmentid",
            "description": "设备 ID"
          },
          "equipmentNo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipmentno",
            "description": "设备编号"
          },
          "equipmentName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipmentname",
            "description": "设备名称"
          },
          "equipmentStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipmentstatus",
            "description": "设备状态"
          },
          "userId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userid",
            "description": "报修用户 ID"
          },
          "userName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "报修用户姓名"
          },
          "damageDescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Damagedescription",
            "description": "损坏说明"
          },
          "damageImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Damageimages",
            "description": "损坏图片地址"
          },
          "repairUserId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repairuserid",
            "description": "维修人员 ID"
          },
          "repairUserName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repairusername",
            "description": "维修人员姓名"
          },
          "assignRemark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignremark",
            "description": "派单备注"
          },
          "assignTime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigntime",
            "description": "派单时间"
          },
          "faultCause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faultcause",
            "description": "故障原因"
          },
          "repairProcess": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repairprocess",
            "description": "维修过程"
          },
          "repairResult": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repairresult",
            "description": "维修结果"
          },
          "beforeImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Beforeimages",
            "description": "维修前图片地址"
          },
          "afterImages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Afterimages",
            "description": "维修后图片地址"
          },
          "completionTime": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completiontime",
            "description": "提交维修完成时间"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "title": "Createtime",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "Updatetime",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "status",
          "id",
          "repairReportId",
          "equipmentId",
          "createTime",
          "updateTime"
        ],
        "title": "RepairOrderOut"
      },
...
      "Result_Page_RepairOrderPageOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_RepairOrderPageOut_"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[Page[RepairOrderPageOut]]"
      },
...
      "Result_RepairOrderActionOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairOrderActionOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[RepairOrderActionOut]"
      },
...
      "Result_RepairOrderCompletionOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairOrderCompletionOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[RepairOrderCompletionOut]"
      },
      "Result_RepairOrderOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairOrderOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[RepairOrderOut]"
      },
...
    },
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}